requirements.txt
# Production
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
PyJWT==2.9.0
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
psycopg2-binary==2.9.9
Pillow==10.4.0

# Development
pytest==8.3.2

# Testing
pytest-flask==1.3.0
